Refact de seguimiento_pilos por report_trackings, seguimiento_grupal por groupal_tracking y pasar seguimiento_grupal a AMD
