+++ managers/permissions_management/permissions_report.php
@@ -1,6 +1,6 @@
 <?php
 require_once('permissions_lib.php');
-require_once('../pilos_tracing/pilos_tracing_lib.php');
+require_once('../pilos_tracking/pilos_tracking_lib.php');
 
 $msg = new stdClass();
