+++ managers/pilos_tracking/tracking_functions.php
@@ -1,5 +1,5 @@
 <?php
-require_once('pilos_tracing_lib.php');
+require_once('pilos_tracking_lib.php');
 
 /*
  * Funciones del módulo seguimientos pilos que se utilizarán en la view. 
@@ -19,6 +19,7 @@ require_once('pilos_tracing_lib.php');
  * @return string
  **/
 function has_tracking($seguimientos){
+    $table="";
     if($seguimientos==""){
      $table.="<p class='text-center'><strong>No existen seguimientos en el periodo seleccionado</strong></p>";
     }else{
