+++ managers/seguimiento_grupal/seguimientogrupal_lib.php
@@ -3,7 +3,7 @@ require_once(dirname(__FILE__).'/../../../../config.php');
 require_once('../MyException.php');
 require_once ('../lib/student_lib.php');
 require_once('../periods_management/periods_lib.php');
-require_once('../pilos_tracing/pilos_tracing_lib.php');
+require_once('../pilos_tracking/pilos_tracking_lib.php');
