+++ templates/groupal_tracking.mustache
@@ -138,5 +138,4 @@
               </div>
             </div>
           </div>
-    <script type="text/javascript" src="../amd/require-setup.js"></script>
-     <script data-main="../amd/src/seguimiento_grupal/main" src="../amd/scripts/require.js"></script>
+
