+++ view/report_trackings.php
@@ -28,12 +28,13 @@
 // Standard GPL and phpdocs
 require_once(__DIR__ . '/../../../config.php');
 require_once($CFG->libdir.'/adminlib.php');
-require_once('../managers/pilos_tracing/tracing_functions.php');
+require_once('../managers/pilos_tracking/tracking_functions.php');
 require_once('../managers/instance_management/instance_lib.php');
 
 
 include('../lib.php');
 include("../classes/output/renderer.php");
+include("../classes/output/report_trackings_page.php");
 
 global $PAGE, $USER;
 
@@ -56,7 +57,7 @@ $contextblock =  context_block::instance($blockid);
 $PAGE->set_context($contextcourse);
 
 
-$url = new moodle_url("/blocks/ases/view/seguimiento_pilos.php",array('courseid' => $courseid, 'instanceid' => $blockid));
+$url = new moodle_url("/blocks/ases/view/report_tackings.php",array('courseid' => $courseid, 'instanceid' => $blockid));
 
 
 //Navigation setup
@@ -110,13 +111,13 @@ if($usernamerole=='monitor_ps'){
     //organiza el select de periodos.
     $table_periods.=get_period_select($periods);
     
-	//Se recupera los estudiantes de un profesional en la instancia y se organiza el array que será transformado en el toogle.
+    //Se recupera los estudiantes de un profesional en la instancia y se organiza el array que será transformado en el toogle.
     $seguimientos = profesionalUser($globalArregloPares,$globalArregloGrupal,$USER->id,$blockid,$userrole,$intervalo_fechas);
     $table.=has_tracking($seguimientos);
 
 }elseif($usernamerole=='sistemas' or $username == "administrador" or $username == "sistemas1008" or $username == "Administrador"){
 
-	//Obtiene los periodos existentes y los roles que contengan "_ps".
+    //Obtiene los periodos existentes y los roles que contengan "_ps".
     $roles = get_rol_ps();
 
     //Obtiene las personas que se encuentran en el último semestre añadido y cuyos roles terminen en "_ps.
@@ -158,13 +159,13 @@ $PAGE->requires->css('/blocks/ases/js/DataTables-1.10.12/css/NewCSSExport/button
 $PAGE->requires->css('/blocks/ases/js/DataTables-1.10.12/css/dataTables.tableTools.css', true);
 $PAGE->requires->css('/blocks/ases/style/sweetalert.css', true);
 $PAGE->requires->css('/blocks/ases/js/select2/css/select2.css', true);
-$PAGE->requires->js_call_amd('block_ases/pilos_tracing_main','init');
+$PAGE->requires->js_call_amd('block_ases/pilos_tracking_main','init');
 $PAGE->set_url($url);
 $PAGE->set_title($title);
 
 $output = $PAGE->get_renderer('block_ases');
 
 echo $output->header();
-$seguimiento_pilos_page = new \block_ases\output\seguimiento_pilos_page($data);
-echo $output->render($seguimiento_pilos_page);
+$report_trackings_page = new \block_ases\output\report_trackings_page($data);
+echo $output->render($report_trackings_page);
 echo $output->footer();
